Add caching of task lists in json file
A problem with this app was that the tasks lists
were taking a long time to load. Since the tasks
lists change rarely, it makes sense to cache
the task lists on the PHP server so that they
can be displayed immediately to the user without
having to wait for the Google API to return all
the task lists. This change stores the task
lists from the last API call to google, in a text file
on the server. This file is loaded first, before
attempting to access the Tasks API for
the new list. When the result comes back from the
server, the user's selection is preserved.